+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- add push_commit method and refactor branch method to branch_or_main(pr [#286])
+
 ## [0.4.0] - 2024-08-17
 
 ### Added
@@ -526,6 +532,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#283]: https://github.com/jerus-org/pcu/pull/283
 [#284]: https://github.com/jerus-org/pcu/pull/284
 [#285]: https://github.com/jerus-org/pcu/pull/285
+[#286]: https://github.com/jerus-org/pcu/pull/286
+[Unreleased]: https://github.com/jerus-org/pcu/compare/v0.4.0...HEAD
 [0.4.0]: https://github.com/jerus-org/pcu/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/jerus-org/pcu/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/jerus-org/pcu/compare/v0.1.26...v0.2.0