+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - add push_commit method and refactor branch method to branch_or_main(pr [#286])
 
+### Fixed
+
+- integrate get_authenticated_remote into push_commit method(pr [#287])
+
 ## [0.4.0] - 2024-08-17
 
 ### Added
@@ -533,6 +537,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#284]: https://github.com/jerus-org/pcu/pull/284
 [#285]: https://github.com/jerus-org/pcu/pull/285
 [#286]: https://github.com/jerus-org/pcu/pull/286
+[#287]: https://github.com/jerus-org/pcu/pull/287
 [Unreleased]: https://github.com/jerus-org/pcu/compare/v0.4.0...HEAD
 [0.4.0]: https://github.com/jerus-org/pcu/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/jerus-org/pcu/compare/v0.2.0...v0.3.0