+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - integrate get_authenticated_remote into push_commit method(pr [#287])
 - add headers and refine error handling(pr [#293])
+- add owner and name fields to repository query(pr [#296])
 
 ## [0.4.0] - 2024-08-17
 
@@ -555,6 +556,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#293]: https://github.com/jerus-org/pcu/pull/293
 [#294]: https://github.com/jerus-org/pcu/pull/294
 [#295]: https://github.com/jerus-org/pcu/pull/295
+[#296]: https://github.com/jerus-org/pcu/pull/296
 [Unreleased]: https://github.com/jerus-org/pcu/compare/v0.4.0...HEAD
 [0.4.0]: https://github.com/jerus-org/pcu/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/jerus-org/pcu/compare/v0.2.0...v0.3.0